requirements.txt
aiogram~=2.22.2
openai~=0.27.2
PyYAML~=6.0.1
rich~=13.5.2